Make changes after sprint 7 first review
