+++ src/kanban/exception/ManagerSaveException.java
@@ -1,0 +1,11 @@
+package kanban.exception;
+
+public class ManagerSaveException extends RuntimeException {
+    public ManagerSaveException(String message) {
+        super(message);
+    }
+
+    public ManagerSaveException(Throwable cause) {
+        super(cause);
+    }
+}
